* use gradle build system for build the project * add github actions for compiling * add task for fat jar
